requirements.txt
fastapi
uvicorn[standard]
pillow
opencv-python
numpy
python-multipart
typing-extensions
torch==2.9.0+cpu
torchvision
onnxruntime==1.24.3

--extra-index-url https://download.pytorch.org/whl/cpu
